server/requirements.txt
djangorestframework==3.12.2